keyvault-requirements.txt
adal==1.2.2
asn1crypto==0.24.0
azure-common==1.1.23
azure-keyvault==1.1.0
azure-nspkg==3.0.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
isodate==0.6.0
msrest==0.6.9
msrestazure==0.6.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
